+++ SOPT_Kakao/SOPT_Kakao/Presentation/Main/Chat/ChatViewController + Extension.swift
@@ -1,0 +1,54 @@
+//
+//  ChatViewController + Extension.swift
+//  SOPT_Kakao
+//
+//  Created by 김인영 on 2022/11/04.
+//
+
+import UIKit
+
+//MARK: - UICollectionViewDelegateFlowLayout
+extension ChatViewController : UICollectionViewDelegateFlowLayout {
+    func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, sizeForItemAt indexPath: IndexPath) -> CGSize {
+        return CGSize(width: UIScreen.main.bounds.width , height: 50)
+    }
+
+    func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, minimumLineSpacingForSectionAt section: Int) -> CGFloat {
+        return 0
+    }
+
+    func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, minimumInteritemSpacingForSectionAt section: Int) -> CGFloat {
+        return 0
+    }
+
+    func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, insetForSectionAt section: Int) -> UIEdgeInsets {
+        return UIEdgeInsets(top: 0, left: 0, bottom: 10, right: 0)
+    }
+    
+    func collectionView(_ collectionView: UICollectionView, viewForSupplementaryElementOfKind kind: String, at indexPath: IndexPath) -> UICollectionReusableView {
+        guard let header = collectionView.dequeueReusableSupplementaryView(ofKind: kind, withReuseIdentifier: ChatHeaderCollectionReusableView.viewIdentifier, for: indexPath) as? ChatHeaderCollectionReusableView else {return UICollectionReusableView()}
+        
+        return header
+    }
+        
+    func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, referenceSizeForHeaderInSection section: Int) -> CGSize {
+        let width = collectionView.frame.width
+        return CGSize(width: width, height: 80)
+    }
+}
+
+// MARK: -UICollectionViewDataSource
+extension ChatViewController: UICollectionViewDataSource {
+    
+    func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
+        return chatList.count
+    }
+    
+    func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
+        guard let chatCell = collectionView.dequeueReusableCell(
+            withReuseIdentifier: ChatCollectionViewCell.identifier, for: indexPath)
+                as? ChatCollectionViewCell else { return UICollectionViewCell() }
+        chatCell.dataBind(model: chatList[indexPath.item])
+        return chatCell
+    }
+}
