+++ SOPT_Kakao/SOPT_Kakao/Presentation/Main/Photo/PhotoCollectionViewController.swift
@@ -40,7 +40,7 @@ class PhotoCollectionViewController: UIViewController {
     }
     
     //MARK: - CollectionVC
-    private lazy var photoCollectionView: UICollectionView = {
+    lazy var photoCollectionView: UICollectionView = {
         let layout = UICollectionViewFlowLayout()
         layout.scrollDirection = .vertical
         
@@ -59,7 +59,6 @@ class PhotoCollectionViewController: UIViewController {
         super.viewDidLoad()
         layout()
         register()
-        
     }
     
     //MARK: -  PhotoList
@@ -107,6 +106,7 @@ class PhotoCollectionViewController: UIViewController {
     }
 }
 
+//MARK: - Extension - Layout
 extension PhotoCollectionViewController {
     
     func layout() {
@@ -142,8 +142,6 @@ extension PhotoCollectionViewController {
             make.bottom.equalToSuperview()
             make.height.equalTo(calculateHeight())
         }
-        
-    
     }
     
     // collectionView의 height 계산
@@ -162,63 +160,8 @@ extension PhotoCollectionViewController {
         return heightCount * photoCellHeight + (heightCount - 1) * photoLineSpacing + photoInset.top + photoInset.bottom
     }
     
+    // MARK: - Register
     private func register() {
         photoCollectionView.register(PhotoCollectionViewCell.self, forCellWithReuseIdentifier: PhotoCollectionViewCell.identifier)
-    }
-}
-
-
-extension PhotoCollectionViewController: UICollectionViewDelegateFlowLayout {
-    
-    func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, sizeForItemAt indexPath: IndexPath) -> CGSize {
-        let screenWidth = UIScreen.main.bounds.width
-        let doubleCellWidth = screenWidth - photoInset.left - photoInset.right - photoInterItemSpacing * 2
-        return CGSize(width: doubleCellWidth / 3, height: photoCellHeight)
-    }
-
-    func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, minimumLineSpacingForSectionAt section: Int) -> CGFloat {
-        return photoLineSpacing
-    }
-
-    func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, minimumInteritemSpacingForSectionAt section: Int) -> CGFloat {
-        return photoInterItemSpacing
-    }
-
-    func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, insetForSectionAt section: Int) -> UIEdgeInsets {
-        return photoInset
-    }
-}
-
-extension PhotoCollectionViewController: UICollectionViewDataSource {
-    func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
-        return photoList.count
-    }
-    
-    func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
-        guard let photoCell = photoCollectionView.dequeueReusableCell(withReuseIdentifier: PhotoCollectionViewCell.identifier, for: indexPath) as? PhotoCollectionViewCell else { return UICollectionViewCell() }
-        photoCell.dataBind(model: photoList[indexPath.item], isSelected: photoCell.isSelected)
-        return photoCell
-    }
-    
-    //didSelect로 돌아가지 않았음..
-//    func collectionView(_ collectionView: UICollectionView, didSelectItemAt indexPath: IndexPath) {
-//        let photoCell = photoCollectionView.cellForItem(at: indexPath) as! PhotoCollectionViewCell
-//        if indexPath.row == 0 {
-//            photoCell.deselectedLayout()
-//        } else {
-//            photoCell.dataBind(model: photoList[indexPath.row], isSelected: photoCell.isSelected )
-//        }
-//        photoCollectionView.reloadData()
-//    }
-
-    func collectionView(_ collectionView: UICollectionView, shouldSelectItemAt indexPath: IndexPath) -> Bool {
-        guard let photoCell = photoCollectionView.cellForItem(at: indexPath) as? PhotoCollectionViewCell else {
-            return true}
-        if photoCell.isSelected {
-            photoCollectionView.deselectItem(at: indexPath, animated: true)
-            return false
-        } else {
-            return true
-        }
     }
 }
